+++ proguardprocessor/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- feat: add telemetry support to proguard symbolicator
+- feat: handle stack frames that don't need symbolication (#106) | @jairo-mendoza
+- feat: add telemetry support to proguard symbolicator (#105) | @wolfgangcodes
 
 ## v0.0.4 [beta] - 2025/08/20